requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
pdfplumber>=0.10.0
faiss-cpu>=1.7.4
numpy>=1.24.0

